% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{adjust_p}
\alias{adjust_p}
\title{Adjust p-values stored in a data.frame using the Holm-Bonferroni correction}
\usage{
adjust_p(df, p_col = "p.value", method = "holm")
}
\arguments{
\item{df}{A data.frame containing a column of p-values to be adjusted}

\item{p_col}{A string defining the name of the column in the input data.frame
that contains the p-values to be adjusted.}

\item{method}{A string defining the name of the p-value adjustment method to
be used. Defautls to Holm's method.}
}
\value{
A data.frame identical to the original, with an added column
  containing the adjusted p-values.
}
\description{
Adjust p-values stored in a data.frame using the Holm-Bonferroni correction
}
