% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{calc_cov_mat}
\alias{calc_cov_mat}
\title{Calculates a covariance matrix from a set of correlation coefficients and
corresponding standard deviations}
\usage{
calc_cov_mat(rho, std_dev, ncond = 2)
}
\arguments{
\item{rho}{Correlation coefficient, or a number between -1 and 1.}

\item{std_dev}{Vector of standard deviations.}

\item{ncond}{How many columns should the covariance matrix have. Default = 2.}
}
\value{
The corresponding covariance matrix.
}
\description{
Calculates a covariance matrix based on a series of correlation coefficients
and standard deviations.
This version assumes a single correlation coefficient and a single standard
deviation for all pairs of measurements.
}
