% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{calc_error_design}
\alias{calc_error_design}
\title{Calculates the type-S and type-M errors in a given power simulation.}
\usage{
calc_error_design(sim_results, alpha = 0.05, rename_cols = TRUE)
}
\arguments{
\item{sim_results}{A matrix with all the simulation results in the columns.}

\item{alpha}{A double between 0 and 1, defining the significance level used in the
simulations Defaults to .05.}

\item{rename_cols}{A boolean. Whether or not to rename the columns. Defaults
to TRUE. At the moment, this function requires this to be set to true.}
}
\value{
A vector with the relevant information about the simulations (the
  simulated effect size, power (unadjusted and adjusted), type S and M error
  rates (unadjusted and adjusted)).
}
\description{
Calculates the type-S (sign) and type-M (exaggeration ratio) errors for a
particular design on a power simulation, for both unadjusted and adjusted
p-values.
}
