% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_sims}
\alias{combine_sims}
\title{Utility function to combine output of power simulations with multiple
parameters.}
\usage{
combine_sims(sim_list)
}
\arguments{
\item{sim_list}{The output of a power simulation.}
}
\value{
A list with all the columns adequately labeled.
}
\description{
This function takes a simulation list output and adds simulation and
condition numbers and outputs a proper data.frame for data exploration
}
