% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{correlation2covariance}
\alias{correlation2covariance}
\title{Generate a covariance matrix from a correlation matrix and corresponding
standard deviations}
\usage{
correlation2covariance(cor_mat, sd_var = 50)
}
\arguments{
\item{cor_mat}{Correlation matrix.}

\item{sd_var}{Vector of standard deviations}
}
\value{
The corresponding covariance matrix.
}
\description{
Generates a covariance matrix based on a user-given correlation matrix.
This version assumes a single standard deviation for all pairs of
measurements.

adapted from: https://stats.stackexchange.com/questions/62850/obtaining-covariance-matrix-from-correlation-matrix
}
