% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{fast_t}
\alias{fast_t}
\title{Fast calculation of a paired t-test}
\usage{
fast_t(data_matrix)
}
\arguments{
\item{data_matrix}{A matrix with two columns upon which the paired t-test
will be computed.}
}
\value{
A list with the same statistics as the base R version, namely:
  t statistic, degrees of freedom, p-value, 95% confidence interval, mean
  difference and its standard error.
}
\description{
Calculates a paired t-test more rapidly than the base R version, making it
suitable for simulations where the same operation needs to be repeated a very
large number of times.

This version only calculates the paired t-test, based on a matrix with two
data columns.
}
