% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matcols2lists}
\alias{matcols2lists}
\title{Breaks down a matrix into a list where each item is an arbitrary number of
columns from the original matrix.}
\usage{
matcols2lists(data_matrix, group_by = 2)
}
\arguments{
\item{data_matrix}{Input matrix.}

\item{group_by}{How many columns are supposed to be in each group that will
become a list item. Defaults to 2}
}
\value{
A list with all the columns from the input matrix rearranged as
  items, each with the user-specified number of columns
}
\description{
This function takes a matrix and breaks it into a user-specified number of
columns, where each group becomes a list item.

For the time being, this is a simple helper function, and does not perform
any checks on the input, so the user needs to make sure the input is correct,
otherwise unpredictable behavior can occur.
}
