% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{organize_output}
\alias{organize_output}
\title{Utility function to organize output of power simulations.}
\usage{
organize_output(simdata_list)
}
\arguments{
\item{simdata_list}{The output of a power simulation..}
}
\value{
A list with all the columns adequately labeled.
}
\description{
This function takes a simulation list output and organizes it for further
processing. It is mostly a convenient way of shortening the data processing
pipeline.
}
