% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_list}
\alias{rbind_list}
\title{Applies rbind() to a list}
\usage{
rbind_list(input_list)
}
\arguments{
\item{input_list}{Input list, generally the product of a pipe context.}
}
\value{
A new data.table or matrix that pulls together all the items from
 the input list.
}
\description{
`rbind_list()` takes a list as an argument and applies `rbind()` to it to
create a data.table or matrix element.

With the introduction of native pipes in base R, this functions allows one
one to pipe a list that is being processed in a pipe context and pass it to
rbind using to `do.call()`
}
\examples{
list(a = rnorm(10), b = rnorm(10)) |>
  lapply(mean) |>
  rbind_list()
}
\seealso{
[rbind()], which this function wraps.
}
