% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_data_from_es}
\alias{simulate_data_from_es}
\title{Function factory to generated a simulation function based on a user-specified
set of effect sizes}
\usage{
simulate_data_from_es(es, base_mu = 600)
}
\arguments{
\item{es}{Vector representing the effects sizes to be simulated in the
generated function}

\item{base_mu}{A double. If one desires to use more realistic numbers in the
simulation. Defaults to 600, a sensible mean reaction time for lexical
decision experiments.}
}
\value{
A function that closes over the effects sizes and can be
  further parametrized by the user.
}
\description{
This is a function factory, i.e., a closure that takes a set of effect sizes
specified by the user and outputs a function that the user can parametrize
in terms of correlations, standard deviations and sample sizes.

The use case in in large scale power simulations.
}
