% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{t_test_paired_data}
\alias{t_test_paired_data}
\title{Convenient wrapper for fast paired t-test calculation}
\usage{
t_test_paired_data(
  paired_data,
  fast = TRUE,
  use_names = TRUE,
  info_out = c("estimate", "statistic", "p.value")
)
}
\arguments{
\item{paired_data}{A matrix with two columns upon which the paired t-test
will be computed.}

\item{fast}{Boolean. Should the fast paired t-test function be used? Defaults
to TRUE.}

\item{use_names}{Boolean. Use existing column names in the output? Defaults
to TRUE. If FALSE, column names are set to NA.}

\item{info_out}{Vector of strings. Names the fields from the t-test function
used in the output. Defaults to "estimate", "statistic" and "p.value".}
}
\value{
A vector with the summary information described in "info_out".
}
\description{
Convenient wrapper for fast paired t-test that outputs a sensible default
subset of information, and deals with the names of the columns.

The use case for this function is in large simulations, to automatically
take care of handling column names and subsetting output.
}
