% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{test_2x2_paired_data}
\alias{test_2x2_paired_data}
\title{Calculates the main effects and the interaction in a 2x2 within-subjects
factorial design}
\usage{
test_2x2_paired_data(sim_results, rename_cols = TRUE)
}
\arguments{
\item{sim_results}{A matrix with four columns upon which the three
paired t-tests will be computed.}

\item{rename_cols}{Boolean. Automatically rename the columns? Defaults
to TRUE, which should give sensible default names.}
}
\value{
A vector with the summary information of the three paired t-tests.
}
\description{
Calculates the main effects and the interaction in a 2x2 within-subjects
factorial design by means of three paired t-tests. The interaction is
equivalent as a paired t-test of the difference between the main effects.

It automatically calculates the adjusted p-values for the three tests, in
case it is needed.

The use case for this function is in large simulations, to automatically
take care of handling column names and subsetting output.
}
