% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_functions.R
\name{vec2onerow_mat}
\alias{vec2onerow_mat}
\title{Transforms a vector into a one row matrix.}
\usage{
vec2onerow_mat(vec)
}
\arguments{
\item{vec}{A vector to be transformed into a one row matrix}
}
\value{
A one row matrix with the contents of vec as columns.
}
\description{
Transforms a vector into a matrix with one row. This is a hacky quickfix
for when a function expects a matrix but instead gets a vector.
}
